CalibreDb: Start moving filters to query params
